requirements.txt
matplotlib==3.5.2
numpy==1.22.4
onnxruntime==1.11.1
pandas==1.4.2
psutil==5.8.0
pyod==1.0.1
PyYAML==6.0
toml==0.10.2
combo==0.1.3
